FlaskProject/requirements.txt
pandas~=2.2.3
Flask~=3.1.0
psycopg2
os
psycopg2-binary~=2.9.10
python-dotenv~=1.0.1